+++ scripts/swebok-bootstrap.sh
@@ -1,0 +1,1 @@
+../swebok-bootstrap.sh